Dictionary
This shows the dictionary, you can put any of the residential statuses
in the textbox and press score dictionary,

i tried to create a dump of the data base, it's suppose to have the
CREATE included. But it doesnt look like a sql script. so if this work,
you can go to the score_CardManager and uncomment  the line in public
Score_CardManager() , so you can test the other buttons.